adjust type in sum function

+++ nanites/model.py
@@ -1,6 +1,6 @@
 
 
-def sum_ints(a: int, b: float) -> int:
+def sum_ints(a: int, b: int) -> int:
     """Placeholder function that sums two integers.
 
     :param a:       Any integer
